Update resolution fix for Editor Play Mode state change
Fixes a VERY annoying issue whereby on entering Play Mode, Unity will run "Validate" and attempt to initialise the Service Manager. Where this is normally required for normal Unity Editing, this instance is then destroyed before "Awake" is called.
This fix "finally" allows the Service Manager to "skip" initialisation if the Unity Editor enters play mode and prevents double initialization.

Does not affect normal "Play" mode as this is an editor only behaviour.

Also fixed a minor grievance where the Service Manager for whatever reason is not initialised in the Editor and causes the inspector to throw a null reference exception (rare issue)